pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu==1.7.4",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
